+++ tcob/src/audio/audio_codecs/AudioCodec_flac.cpp
@@ -62,10 +62,10 @@ auto flac_decoder::open() -> std::optional<buffer::information>
     return std::nullopt;
 }
 
-auto flac_decoder::decode(std::span<f32> outputSamples) -> i32
+auto flac_decoder::decode(std::span<f32> outputSamples) -> isize
 {
     u64 const wantRead {outputSamples.size() / static_cast<u32>(_info.Specs.Channels)};
-    return static_cast<i32>(drflac_read_pcm_frames_f32(_flac, wantRead, outputSamples.data()) * _info.Specs.Channels);
+    return static_cast<isize>(drflac_read_pcm_frames_f32(_flac, wantRead, outputSamples.data()) * _info.Specs.Channels);
 }
 
 }
